requirements.txt
python-telegram-bot==20.3
SQLAlchemy==2.0.36
python-dotenv==1.0.0
